requirements.txt
Django==2.2.28
django-ckeditor==6.1.0
django-modeltranslation==0.17.3
ldap3==2.5.1
requests==2.26.0
stripe==2.48.0
django-phonenumber-field==5.2.0
mysqlclient==2.0.3
Pillow==8.3.1
django-python3-ldap==0.12.0
uWSGI==2.0.19.1
redis==3.5.3
rq==1.13.0
rq-scheduler==0.13.0
django-rq==2.7.0
django-rq-scheduler==1.1.3
python-logstash>=0.4.6
python-dateutil==2.8.2
coverage==5.5
django_coverage_plugin==2.0.0
django-debug-toolbar==3.2.1
django-redis==5.0.0
phonenumbers==8.12.27
PyYAML==5.4.1
irclog2html==2.17.3
pylint==2.9.4
pylint-django==2.4.4
pyasn1==0.4.8
